moved game button logic to sperate class/object
